fix(codex): avoid OpenAI provider id init cycle
