+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "openclaw",
-  "version": "2026.5.26-openclaw.1",
+  "version": "2026.5.26-openclaw.2",
   "description": "Multi-channel AI gateway with extensible messaging integrations",
   "keywords": [],
   "homepage": "https://github.com/openclaw/openclaw#readme",
