+++ src/agents/openai-codex-routing.ts
@@ -1,82 +1,26 @@
 import type { OpenClawConfig } from "../config/types.openclaw.js";
-import { normalizeOptionalLowercaseString } from "../shared/string-coerce.js";
 import { normalizeEmbeddedAgentRuntime } from "./pi-embedded-runner/runtime.js";
 import { resolveProviderIdForAuth } from "./provider-auth-aliases.js";
 import { findNormalizedProviderValue, normalizeProviderId } from "./provider-id.js";
-
-export const OPENAI_PROVIDER_ID = "openai";
-export const OPENAI_CODEX_PROVIDER_ID = "openai-codex";
-
-function isOfficialOpenAIBaseUrl(baseUrl: unknown): boolean {
-  if (typeof baseUrl !== "string" || !baseUrl.trim()) {
-    return true;
-  }
-  try {
-    const url = new URL(baseUrl.trim());
-    return (
-      url.protocol === "https:" &&
-      url.hostname.toLowerCase() === "api.openai.com" &&
-      (url.pathname === "" ||
-        url.pathname === "/" ||
-        url.pathname === "/v1" ||
-        url.pathname === "/v1/")
-    );
-  } catch {
-    return false;
-  }
-}
-
-function openAIProviderUsesCustomBaseUrl(config: OpenClawConfig | undefined): boolean {
-  return !isOfficialOpenAIBaseUrl(config?.models?.providers?.openai?.baseUrl);
-}
-
-export function isOpenAIProvider(provider: string | undefined): boolean {
-  return normalizeProviderId(provider ?? "") === OPENAI_PROVIDER_ID;
-}
-
-export function isOpenAICodexProvider(provider: string | undefined): boolean {
-  return normalizeProviderId(provider ?? "") === OPENAI_CODEX_PROVIDER_ID;
-}
-
-export function openAIProviderUsesCodexRuntimeByDefault(params: {
-  provider?: string;
-  config?: OpenClawConfig;
-}): boolean {
-  return isOpenAIProvider(params.provider) && !openAIProviderUsesCustomBaseUrl(params.config);
-}
-
-export function parseModelRefProvider(value: unknown): string | undefined {
-  if (typeof value !== "string") {
-    return undefined;
-  }
-  const slashIndex = value.trim().indexOf("/");
-  if (slashIndex <= 0) {
-    return undefined;
-  }
-  return normalizeProviderId(value.trim().slice(0, slashIndex));
-}
-
-export function modelRefUsesOpenAIProvider(value: unknown): boolean {
-  return parseModelRefProvider(value) === OPENAI_PROVIDER_ID;
-}
-
-export function modelSelectionShouldEnsureCodexPlugin(params: {
-  model?: string;
-  config?: OpenClawConfig;
-}): boolean {
-  const provider = parseModelRefProvider(params.model);
-  if (provider === OPENAI_CODEX_PROVIDER_ID) {
-    return true;
-  }
-  return provider === OPENAI_PROVIDER_ID && !openAIProviderUsesCustomBaseUrl(params.config);
-}
-
-export function hasOpenAICodexAuthProfileOverride(value: unknown): boolean {
-  return (
-    typeof value === "string" &&
-    normalizeOptionalLowercaseString(value)?.startsWith(`${OPENAI_CODEX_PROVIDER_ID}:`) === true
-  );
-}
+export {
+  OPENAI_CODEX_PROVIDER_ID,
+  OPENAI_PROVIDER_ID,
+  hasOpenAICodexAuthProfileOverride,
+  isOpenAICodexProvider,
+  isOpenAIProvider,
+  modelRefUsesOpenAIProvider,
+  modelSelectionShouldEnsureCodexPlugin,
+  openAIProviderUsesCodexRuntimeByDefault,
+  parseModelRefProvider,
+  resolveContextConfigProviderForRuntime,
+} from "./openai-codex-runtime-ids.js";
+import {
+  OPENAI_CODEX_PROVIDER_ID,
+  OPENAI_PROVIDER_ID,
+  hasOpenAICodexAuthProfileOverride,
+  isOpenAIProvider,
+  openAIProviderUsesCodexRuntimeByDefault,
+} from "./openai-codex-runtime-ids.js";
 
 function configuredOpenAIAuthOrderStartsWithCodexProfile(config: OpenClawConfig | undefined) {
   if (!openAIProviderUsesCodexRuntimeByDefault({ provider: OPENAI_PROVIDER_ID, config })) {
@@ -192,16 +136,4 @@ export function resolveSelectedOpenAIPiRuntimeProvider(params: {
     configuredOpenAIAuthOrderStartsWithCodexProfile(params.config)
     ? OPENAI_CODEX_PROVIDER_ID
     : params.provider;
-}
-
-export function resolveContextConfigProviderForRuntime(params: {
-  provider: string;
-  runtimeId?: string;
-}): string {
-  const provider = normalizeProviderId(params.provider);
-  const runtimeId = normalizeEmbeddedAgentRuntime(params.runtimeId);
-  if (provider === OPENAI_PROVIDER_ID && runtimeId === "codex") {
-    return OPENAI_CODEX_PROVIDER_ID;
-  }
-  return params.provider;
 }
